requirements.txt
psycopg2
kafka-python>=2.0.1